+++ src/main/java/j_parser/types/constants/ConstantMethodRef.java
@@ -5,20 +5,20 @@ import j_parser.utils.ConstantTag;
 
 public final class ConstantMethodRef implements Constant{
     private final int TAG = ConstantTag.METHODREF.TAG;
-    private final short CLASS_INDEX;
-    private final short NT_INDEX;
+    private final int CLASS_INDEX;
+    private final int NT_INDEX;
     private final String COMMON_NAME = ConstantTag.METHODREF.name();
 
-    public ConstantMethodRef(short classIndex, short NTIndex){
+    public ConstantMethodRef(int classIndex, int NTIndex){
         this.CLASS_INDEX = classIndex;
         this.NT_INDEX = NTIndex;
     }
     
-    public short getClassIndex(){
+    public int getClassIndex(){
         return this.CLASS_INDEX;
     }
 
-    public short getNTIndex(){
+    public int getNTIndex(){
         return this.NT_INDEX;
     }
