+++ src/main/java/j_parser/types/constants/ConstantModule.java
@@ -5,14 +5,14 @@ import j_parser.interfaces.Constant;
 
 final public class ConstantModule implements Constant{
     private final int TAG = ConstantTag.MODULE.TAG;
-    private final short NAME_INDEX; 
+    private final int NAME_INDEX; 
     private final String COMMON_NAME = ConstantTag.MODULE.name();
 
-    public ConstantModule(short nameIndex){
+    public ConstantModule(int nameIndex){
         this.NAME_INDEX = nameIndex;
     }
     
-    public short getNameIndex(){
+    public int getNameIndex(){
         return this.NAME_INDEX;
     }
